Revert "signalId from Integer to Long"
This reverts commit 054120affec971d2413c18cbe8ce551c63918aac.

+++ backend/src/main/java/com/sentinel/backend/controller/SignalController.java
@@ -86,7 +86,7 @@ public class SignalController {
 
     @GetMapping("/{signalId}/poll/results")
     public ResponseEntity<ApiResponse<PollResultDTO>> results(
-            @PathVariable Long signalId) {
+            @PathVariable Integer signalId) {
 
         long start = System.currentTimeMillis();
 
@@ -131,7 +131,7 @@ public class SignalController {
 
     @DeleteMapping("/{signalId}")
     public ResponseEntity<ApiResponse<Void>> delete(
-            @PathVariable Long signalId) {
+            @PathVariable Integer signalId) {
 
         long start = System.currentTimeMillis();
